+++ src/main/java/de/jeff_media/angelchest/hooks/ItemsAdderHook.java
@@ -1,0 +1,7 @@
+package de.jeff_media.angelchest.hooks;
+
+public class ItemsAdderHook {
+
+
+
+}